[CoreBundle] adds widgets grid
