--- main/core/Resources/modules/modals/resource/store/index.js
@@ -1,3 +1,0 @@
-/**
- * Created by Corum on 12/06/2018.
- */
