--- main/core/Resources/modules/widget/actions.js
@@ -1,4 +1,0 @@
-
-// todo : grab some logic from home tool
-
-export const actions = {}
