+++ main/core/Resources/modules/widget/editor/modals/creation/store/index.js
@@ -1,0 +1,18 @@
+/**
+ * Widget creation store.
+ */
+
+import {registry} from '#/main/app/store/registry'
+
+import {actions}   from '#/main/core/widget/editor/modals/creation/store/actions'
+import {reducer}   from '#/main/core/widget/editor/modals/creation/store/reducer'
+import {selectors} from '#/main/core/widget/editor/modals/creation/store/selectors'
+
+// append the reducer to the store
+registry.add(selectors.STORE_NAME, reducer)
+
+// export store module
+export {
+  actions,
+  selectors
+}
