+++ main/core/Resources/modules/widget/editor/modals/creation/store/selectors.js
