+++ main/core/Resources/modules/widget/editor/modals/parameters/store/reducer.js
@@ -1,0 +1,12 @@
+import {makeFormReducer} from '#/main/core/data/form/reducer'
+
+import {Widget as WidgetTypes} from '#/main/core/widget/prop-types'
+import {selectors} from '#/main/core/widget/editor/modals/parameters/store/selectors'
+
+const reducer = makeFormReducer(selectors.STORE_NAME, {
+  data: WidgetTypes.defaultProps
+})
+
+export {
+  reducer
+}
