--- main/core/Resources/modules/widget/modals/parameters/store/actions.js
@@ -1,16 +1,0 @@
-import {actions as formActions} from '#/main/core/data/form/actions'
-
-import {selectors} from '#/main/core/widget/modals/parameters/store'
-
-// action names
-
-
-// action creators
-export const actions = {}
-
-/**
- * Updates the widget configuration.
- */
-actions.save = () => formActions.saveForm(selectors.FORM_NAME, ['claro_resource_action', {
-
-}])
