--- main/core/Resources/modules/widget/modals/parameters/store/selectors.js
@@ -1,10 +1,0 @@
-import {select as formSelectors} from '#/main/core/data/form/selectors'
-
-const STORE_NAME = 'widgetParameters'
-
-const saveEnabled = (state) => formSelectors.saveEnabled(formSelectors.form(state, STORE_NAME))
-
-export const selectors = {
-  STORE_NAME,
-  saveEnabled
-}
