+++ main/core/Resources/modules/widget/types/simple/index.js
@@ -1,6 +1,11 @@
 import {makeReducer} from '#/main/core/scaffolding/reducer'
 
 import {SimpleWidget} from '#/main/core/widget/types/simple/components/widget'
+import {SimpleWidgetEditor} from '#/main/core/widget/types/simple/components/editor'
+
+export const Parameters = (context, parameters) => ({
+  component: SimpleWidgetEditor
+})
 
 /**
  * Simple widget application.
